requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.7
django-appconf==1.0.5
django-compressor==4.3.1
django-crispy-forms==2.0
django-cursor-pagination==0.2.1
django-sass-processor==1.2.2
djangorestframework==3.14.0
Faker==17.6.0
libsass==0.22.0
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
